[FEATURE] Use new build & sync tasks in deploy task
Use the new buildapp and sync tasks in the deploy task.
This commit replaces the previous deploy method of
syncing to the target stage and then fetching all dependencies
with the new approach to first build the app locally and
then sync the result → therefore recommended to sync the
composer vendor folder as well.