[FEATURE] Add task to view stage URLs
Add a task to open the project URL for a given stage
in the default browser.

The stage needs to have a origin parameter set in the properties file.

The command to view the project on the test stage would be
`./robo.phar view -s test`, which opens »https://test.example.com«
in a browser.