[FEATURE] Open browser after deployment
Whenever a deployment to a stage is done, then the script
may open a browser with the URL of the project on that stage.

Exception: This is not done on the local stage, because
this could get annoying very soon.

Add parameter »settings.view.open-browser-after-deployment«
to enable/disable the feature completely for the project.